move over get_tournament_details file
