+++ chessdotcom/endpoints/__init__.py
@@ -19,6 +19,7 @@ from .team_match import get_team_match
 from .team_match_board import get_team_match_board
 from .team_match_live import get_team_match_live
 from .team_match_live_board import get_team_match_live_board
+from .tournament_details import get_tournament_details
 
 
 @Client.endpoint
@@ -80,24 +81,6 @@ def get_player_team_matches(
         uri=f"/player/{username}/matches",
         tts=tts,
         top_level_attribute="matches",
-        request_options=request_options,
-    )
-
-
-@Client.endpoint
-def get_tournament_details(
-    url_id: str, tts=0, **request_options
-) -> ChessDotComResponse:
-    """
-    :param url_id: URL for the club's web page on www.chess.com.
-    :param tts: the time the client will wait before making the first request.
-    :returns: ``ChessDotComResponse`` object containing details about a daily,
-                live and arena tournament.
-    """
-    return Resource(
-        uri=f"/tournament/{url_id}",
-        tts=tts,
-        top_level_attribute="tournament",
         request_options=request_options,
     )
